[GPRS] BSSGP Flow Control implementation
This implements the per-BVC and per-MS flow control as specified in
Section 8.2 of TS 08.18.

In fact, the per-MS flow control is a bit ugly, as BSSGP so far does not keep
any per-MS (per-TLLI) state structures.  So rather than adding them, we
piggy-pack the per-MS flow control onto the sgsn_mm_ctx structure.  This
in turn means that the GMM code needs to call into the BSSGP code to set
the BVC-specific default values for the per-MS flow control when it allocates
a MM context.